refactor: STT and TTS service, slightly imporve of performance
